+++ tests/precision_known_enum_val_failure_boundary_clean.adb
@@ -1,0 +1,6 @@
+procedure Precision_Known_Enum_Val_Failure_Boundary_Clean is
+   type Color is (Red, Green, Blue);
+   C : Color;
+begin
+   C := Color'Val (2);
+end Precision_Known_Enum_Val_Failure_Boundary_Clean;