+++ tests/precision_known_enum_val_failure_character_clean.adb
@@ -1,0 +1,5 @@
+procedure Precision_Known_Enum_Val_Failure_Character_Clean is
+   C : Character;
+begin
+   C := Character'Val (300);
+end Precision_Known_Enum_Val_Failure_Character_Clean;