+++ tests/precision_known_enum_val_failure_dynamic_clean.adb
@@ -1,0 +1,6 @@
+procedure Precision_Known_Enum_Val_Failure_Dynamic_Clean (N : Integer) is
+   type Color is (Red, Green, Blue);
+   C : Color;
+begin
+   C := Color'Val (N);
+end Precision_Known_Enum_Val_Failure_Dynamic_Clean;