+++ tests/precision_known_enum_val_failure_finding.adb
@@ -1,0 +1,6 @@
+procedure Precision_Known_Enum_Val_Failure_Finding is
+   type Color is (Red, Green, Blue);
+   C : Color;
+begin
+   C := Color'Val (5);
+end Precision_Known_Enum_Val_Failure_Finding;